Fix mor compilation warning and fix few compilation configurations
git-svn-id: http://rdinnovation.onf.fr/svn/computreev6@309 b135c390-378b-4ba5-bdcd-e4540112eb5a

+++ library/ctlibio/exporters/grid3d/pb_grid3dastableexporter.cpp
@@ -103,20 +103,19 @@ bool PB_Grid3DAsTableExporter::exportItem(const CT_AbstractGrid3D* item, const Q
     QTextStream stream(&file);
 
     // write header
-    size_t xdim = item->xdim();
-    size_t ydim = item->ydim();
-    size_t zdim = item->zdim();
+    int xdim = item->xdim();
+    int ydim = item->ydim();
+    int zdim = item->zdim();
 
     stream << "Xcorner\tYcorner\tZcorner\tXcenter\tYcenter\tZcenter\tColX\tRowY\tLevZ\tValue\n";
 
     // write data
-    for (size_t xx = 0 ; xx < xdim ; xx++)
+    for (int xx = 0 ; xx < xdim ; xx++)
     {
-        for (size_t yy = 0 ; yy < ydim ; yy++)
+        for (int yy = 0 ; yy < ydim ; yy++)
         {
-            for (size_t zz = 0 ; zz < zdim ; zz++)
+            for (int zz = 0 ; zz < zdim ; zz++)
             {
-
                 size_t sIndex = 0;
                 item->index(xx, yy, zz, sIndex);
